Make ProtoEvo runnable on modern JDKs and add usability fixes
Build / run
- Lower sourceCompatibility from 23 to 21 (matches widely-installable LTS).
- Strip JVM args removed in JDK 17+ (--add-modules jdk.incubator.foreign,
  --illegal-access=permit) which prevented launch on any modern JVM.
- Fix three GLSL fragment shaders that mixed int * float (rejected by strict
  drivers, crashing on first sim creation): shaders/pause, shaders/chemical,
  shaders/blur. Promote `int filter_radius` to float.
- Add run.bat convenience script that pins JDK 21 + Gradle on PATH.

Stability
- PlotGrid: setMajorTicks / setMinorTicks were called with `range/5`, which
  is 0 when only one stat snapshot exists. The drawAxis loops then ran
  `for (x = 0; x <= xMax; x += 0)` — an infinite loop spinning inside
  GlyphLayout.calculateWidths and freezing the entire window. Sanitize
  non-finite/<=0 ticks at the setter and add a 200-iteration safety cap
  in the draw loops.
- Simulation.run: replace `if (paused) continue;` busy-wait with a 20ms
  sleep. Pinned a CPU core whenever paused, starving the render thread
  enough that opening a heavy modal screen could mark the LWJGL window
  "not responding".
- REPL.run: handle null line from BufferedReader (EOF / no console). The
  old code NPE'd on line.equals("\n") and tight-looped printing the same
  exception forever when launched without a console.

Time controls
- Simulation.update: when timeDilation > 1, run that many physics substeps
  per render frame at the original safe dt instead of multiplying dt
  directly. Lets the sim run N× faster without breaking physics; capped
  at 64 substeps/frame as a safety. Fractional values use a partial last
  step.
- New hotkeys in SimulationKeyboardControls:
    `]`  -> 2× faster (steps 0.0625 ... 64)
    `[`  -> 2× slower
    `\`  -> reset to 1.0×
  Current speed printed to console.

Performance toggles
- EnvironmentRenderer: add `renderChemicals` and `renderShadows` flags
  with public getters/setters; gate the chemical-field overlay and the
  baked-shadow LightRenderer pass on them. Both are screen-filling
  fragment passes and dominate GPU time.
- New SimulationScreen.toggleLowDetailMode() flips both flags together.
- Hotkeys: F10 toggles low-detail (both), F11 toggles chemicals only.
- New top-bar button (left side) to toggle low-detail mode.

Neural-network mutation fix (high-impact, real bug)
- SynapseGene.cloneWithMutation: previous behavior fully re-rolled weights
  to uniform [-1, 1] on every mutation, making fine-tuning impossible —
  every mutation was a discontinuous jump, so a population could never
  converge on good weights. Switch to standard NEAT-style 90% small
  Gaussian perturb (sigma=0.15) + 10% full re-roll, with weights bounded
  to [-4, 4]. Lets evolved networks actually improve over generations.
- GeneExpressionFunction.tick: remove redundant second setGRNInputs() call
  after the network tick — it was wasted work and immediately overwritten
  on the next tick anyway.

Genetic clusters viewer
- Implement NetworkGenome.distance() (was a stub returning 0). Standard
  NEAT compatibility distance: c1*E/N + c2*D/N + c3*Wbar from excess /
  disjoint genes and matching-gene weight diff.
- New GeneticClustersScreen: snapshot of living protozoa grouped by
  greedy speciation against the distance threshold. Per cluster shows
  population, avg generation, avg energy, avg health, avg age, and a
  color swatch (mean cell color). Capped at 250 cells / 40 clusters
  / 30 displayed for bounded work. Reachable via a top-bar button on
  the simulation screen (right side, multicell icon).

+++ assets/shaders/blur/fragment.glsl
@@ -42,9 +42,9 @@ void main(){
 //            filter_color += 0.25 * weight[i] * texture2D(u_sample2D, v_texCoord0 - vec2(offset[j], 0.0) / u_resolution) * v_color;
 //        }
 //    }
-    float skip = 1;
+    float skip = 1.0;
     vec4 filter_color = vec4(0.);
-    int filter_radius = 3; // int(u_blurRadius);
+    float filter_radius = 3.0; // int(u_blurRadius);
     float sd = filter_radius * skip;
     float k = 1.0 / (2.0 * sd * sd);
     float sum = 0.0;
